Crud conta rodando na view

+++ src/main/java/com/ufc/br/mvp/service/UsuarioService.java
@@ -12,7 +12,7 @@ public class UsuarioService {
 	@Autowired
 	private UsuarioRepository repository;
 	
-	public Usuario find(int id) {
+	public Usuario find(Long id) {
 		return repository.getOne(id);
 	}
 }
